requirements.txt
beautifulsoup4>=4.9.3
PyMuPDF
PyPDF2
google-generativeai>=0.3.1
numpy>=1.19.5
pandas>=1.1.5
pdf2image>=1.14.0
pillow>=8.1.0
pymupdf>=1.22.5
PyYAML>=6.0
requests>=2.25.1
selenium>=4.10.0
webdriver-manager>=3.8.5
tqdm>=4.64.1
jupyter>=1.0.0
ipywidgets>=8.0.0
scikit-learn
nltk
imagehash
